move polygon count test to widgets regression testfile, and check the only returned polygon is the expected one
